+++ lib/log/logtest/logtest_test.go
@@ -18,5 +18,7 @@ func TestExport(t *testing.T) {
 	assert.Len(t, logs, 1)
 	assert.Equal(t, "TestExport", logs[0].Scope)
 	assert.Equal(t, "hello world", logs[0].Message)
+
+	// In dev mode, attributes are not added
 	assert.Equal(t, map[string]any{"key": "value"}, logs[0].Fields)
 }